Make sure combined imports maintain the same namespace
