+++ test/test_bind_names_python33.py
@@ -1,57 +1,8 @@
-import ast
 import sys
 
 import pytest
 
-from python_minifier.rename import add_namespace, resolve_names
-from python_minifier.rename.bind_names import bind_names
-from python_minifier.rename.util import iter_child_namespaces
-from python_minifier.util import is_ast_node
-
-
-def assert_namespace_tree(source, expected_tree):
-    tree = ast.parse(source)
-
-    add_namespace(tree)
-    bind_names(tree)
-    resolve_names(tree)
-
-    actual = print_namespace(tree)
-
-    print(actual)
-    assert actual.strip() == expected_tree.strip()
-
-
-def print_namespace(namespace, indent=''):
-    s = ''
-
-    if not indent:
-        s += '\n'
-
-    def namespace_name(node):
-        if is_ast_node(node, (ast.FunctionDef, 'AsyncFunctionDef')):
-            return 'Function ' + node.name
-        elif isinstance(node, ast.ClassDef):
-            return 'Class ' + node.name
-        else:
-            return namespace.__class__.__name__
-
-    s += indent + '+ ' + namespace_name(namespace) + '\n'
-
-    for name in sorted(namespace.global_names):
-        s += indent + '  - global ' + name + '\n'
-
-    for name in sorted(namespace.nonlocal_names):
-        s += indent + '  - nonlocal ' + name + '\n'
-
-    for binding in sorted(namespace.bindings, key=lambda b: b.name):
-        s += indent + '  - ' + repr(binding) + '\n'
-
-    for child in iter_child_namespaces(namespace):
-        s += print_namespace(child, indent=indent + '  ')
-
-    return s
-
+from helpers import assert_namespace_tree
 
 def test_module_namespace():
     if sys.version_info < (3, 3) or sys.version_info > (3, 4):
